major changes

+++ src/com/singhroshan1999/SpecialPurposeRegisters.java
@@ -2,7 +2,7 @@ package com.singhroshan1999;
 
 class SpecialPurposeRegisters {
     private static short _SP,_PC,_INC,ADDBUF,DATABUFF;
-    private static byte _IR;
+    private static int _IR;
     private static boolean[] _flag= new boolean[8];
     // TODO : special purpose register
     static void  PC(short tbyte){
@@ -15,7 +15,7 @@ class SpecialPurposeRegisters {
     static void  IR(byte tbyte){
         SpecialPurposeRegisters._IR = tbyte;
     }
-    static byte IR(){
+    static int IR(){
         return SpecialPurposeRegisters._IR;
     }
